+++ backend/ecosystem.config.js
@@ -13,6 +13,7 @@ module.exports = {
       NODE_ENV: 'production',
       HOST: '127.0.0.1',
       PORT: 3004,
+      DB_NAME: 'arthistory',
     },
     time: true,
   }],